MonadLog: update format
This format matches the hslogger format. ZonedTime is easier to read in
dev, and most production environments are likely to be
UTC. Additionally, sub-second resolution is probably excessive for our
needs.

Bump the version of `time` rather than try to deal with any version
using conditional compilation or anything like that.