Remove dependency on `canteven-config`. ...
canteven-config brings in some behavioral constraints to a program
that go well beyond any appropriate scope for this library. The
FromJSON instance of `LoggingConfig` should be enough to make this
library easy to use and setup, using canteven-config or not.

The most problematic "behavioral constraint" pulled in is the inability
for a program to define its own command-line arguments.